style: apply cargo clippy --fix across the workspace
Machine-applicable clippy fixes in pir-sdk-client, runtime (lib +
binaries), pir-runtime-core, and block_reader: needless borrows,
useless `.into()` conversions, `ok_or_else` → `ok_or` for cheap
values, manual `div_ceil` / `is_multiple_of` / range `contains`,
single-char `push`, redundant casts, unused `mut` and loop vars.
Semantics-preserving; `cargo check` is clean.

The build/ crate is excluded: `cargo clippy --fix` mis-removes
imports shared across its ~33 binaries and breaks the build. Its
lints need a manual pass.

+++ runtime/src/hint_pool.rs
@@ -453,8 +453,8 @@ fn compute_and_serialize_hint_frame(
     frame.extend_from_slice(&frame_payload_len.to_le_bytes());
     frame.push(RESP_HARMONY_HINTS);
     frame.push(group_id as u8);
-    frame.extend_from_slice(&(padded_n as u32).to_le_bytes());
-    frame.extend_from_slice(&(t_val as u32).to_le_bytes());
+    frame.extend_from_slice(&padded_n.to_le_bytes());
+    frame.extend_from_slice(&t_val.to_le_bytes());
     frame.extend_from_slice(&(m as u32).to_le_bytes());
     frame.extend_from_slice(&flat);
     frame
